Add the upstream xf86-video-dummy as of November 25, 2018.
This gives x11spice access to a modern virtual/offscreen
frame buffer driver for the Xorg X server.  We will then
make slight revisions to make it more suitable for use with Spice.

The overall result should be a more functional stack,
as compared to the existing xf86-video-qxl driver, as we gain
a wide range of Mesa features for free.  It also simplifies
the overall process.

This is from https://gitlab.freedesktop.org/xorg/driver/xf86-video-dummy
as of 850c0516.