Fix hang when launching with invalid config
When launching with no config file or arguments, the "You must provide
some authentication method" dialog appears. Clicking the Close button
would hang the application and it would never quit.

s->draw_command_in_progress was uninitialized, and flush_and_lock()
would get stuck in an infinite loop waiting for it to be 0.

Initialize the entire session struct to 0.

Signed-off-by: Brendan Shanks <bshanks@codeweavers.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>