Provide an attempted optimization when damage reports appear wrong.
With compositing window managers like mutter, when used with the
spice-video-dummy, you can get damage reports that are the whole screen.

Essentially, you just get an indication that the screen has changed,
but no sense of what has changed.

This change detects that behavior and stops trusting those damage
reports, and instead increases the scan frequency to compensate.

Acked-by: Frediano Ziglio <fziglio@redhat.com>