Add a full-screen-fps option.
This option will bypass XDamage and periodic scanning and simply
transmit the whole screen to the spice server at the specified rate.

This will allow the streaming logic to kick in and hopefully for
video codecs to optimize the transmission.  This is particularly
useful for applications that often change the whole screen.

Acked-by: Frediano Ziglio <fziglio@redhat.com>