Fix compliation on gcc 4.X.
gcc 4.x warns if you use a { 0 } initialization construct
for a structure with an initial member that is also a struct.

The { } construct is simpler and appears to work on a wider
range of gcc versions.

This is a correction to fdfdf1107be100b983de1bff4beee8e6360f670b