Bug fix - send the last few lines of a scanned box when rounding badly.
If the number of vertical lines results in an odd dividend, we were
truncating the computation and not sending the last few lines.

This fixes that by first avoiding early division which would truncate
values, and by then rounding up our width and height calculations.

This fixes the bug in the tests introduced in the previous patch.