Add a new test for tall screens; draw outside of scanline range.
This exposes a bug in the current code base.  That is, we
do not scan beyond the 32 scan lines currently given.

That means if we have a screen taller than 1024 pixels, we will
not scan every block fully, and so we can miss changes.

This test exposes that issue, and currently fails.