Perform region checking only for region related scan reports.
This bug became noticable after cb122392, in that we routinely
send a new scan report type, which should not affect region
tracking.

Investigating this revealed that this code was also buggy
in that the region code could cause non region related
scan reports to be discarded.  The code was refactored
to avoid that case.

Acked-by: Frediano Ziglio <fziglio@redhat.com>